Gerenciamento de múltiplos blocos e refatoração do main
Refatora o código para suportar uma lista de blocos, removendo a função de desenho de bloco único e adicionando funções para adicionar e desenhar múltiplos blocos. O main agora utiliza uma ListaDeBlocos e desenha todos os blocos adicionados.

+++ qdblocos.h
@@ -15,7 +15,7 @@ typedef struct {
 
 typedef struct {
     Bloco blocos[_blocos_tam_max_];
-	int quantidade;
+    uint16_t quantidade;
 } ListaDeBlocos;
